refactor(gateway): split backend client tests
